remove unittest2 -- bumps minimum version supported to 2.7/3.5
This is because there's yet another breaking change in python3.10 around
the name of collections.MutableMapping to
collections.abc.MutableMapping.  This means that unittest2 doesn't work
any longer, and that was possibly needed to support pre-2.7/3.5.

Since unittest2 seems to no longer be needed and those versions of
python are quite old by now, we should drop support for it